feat: unify Get Support dialog across all portals
