+++ src/app/patient/layout.tsx
@@ -6,11 +6,13 @@ import { useAuth } from '@/context/AuthContext';
 import PatientSidebar from '@/components/patient/PatientSidebar';
 import PatientTopbar from '@/components/patient/PatientTopbar';
 import LoadingSpinner from '@/components/shared/LoadingSpinner';
+import SupportBot from '@/components/shared/SupportBot';
 
 export default function PatientLayout({ children }: { children: React.ReactNode }) {
   const { user, loading } = useAuth();
   const router = useRouter();
   const [sidebarOpen, setSidebarOpen] = useState(false);
+  const [supportOpen, setSupportOpen] = useState(false);
 
   useEffect(() => {
     if (!loading && (!user || user.role !== 'PATIENT')) {
@@ -37,11 +39,16 @@ export default function PatientLayout({ children }: { children: React.ReactNode 
           onClick={() => setSidebarOpen(false)}
         />
     )}
-      <PatientSidebar open={sidebarOpen} onClose={() => setSidebarOpen(false)} />
+      <PatientSidebar open={sidebarOpen} onClose={() => setSidebarOpen(false)} onOpenSupport={() => setSupportOpen(true)} />
     <div className="flex-1 lg:ml-64 min-w-0">
       <PatientTopbar onMenuClick={() => setSidebarOpen(true)} />
       <main className="p-4 lg:p-6">{children}</main>
     </div>
+    <SupportBot
+      open={supportOpen}
+      onOpen={() => setSupportOpen(true)}
+      onClose={() => setSupportOpen(false)}
+    />
   </div>
 );
 }
