+++ src/app/super-admin/layout.tsx
@@ -6,11 +6,13 @@ import { useAuth } from '@/context/AuthContext';
 import SuperAdminSidebar from '@/components/super-admin/SuperAdminSidebar';
 import SuperAdminTopbar from '@/components/super-admin/SuperAdminTopbar';
 import LoadingSpinner from '@/components/shared/LoadingSpinner';
+import SupportBot from '@/components/shared/SupportBot';
 
 export default function SuperAdminLayout({ children }: { children: React.ReactNode }) {
   const { user, loading } = useAuth();
   const router = useRouter();
   const [sidebarOpen, setSidebarOpen] = useState(false);
+  const [supportOpen, setSupportOpen] = useState(false);
 
   useEffect(() => {
     if (!loading && (!user || user.role !== 'SUPER_ADMIN')) {
@@ -31,7 +33,7 @@ export default function SuperAdminLayout({ children }: { children: React.ReactNo
       {sidebarOpen && (
         <div className="fixed inset-0 bg-black/50 z-30 lg:hidden" onClick={() => setSidebarOpen(false)} />
       )}
-      <SuperAdminSidebar open={sidebarOpen} onClose={() => setSidebarOpen(false)} />
+      <SuperAdminSidebar open={sidebarOpen} onClose={() => setSidebarOpen(false)} onOpenSupport={() => setSupportOpen(true)} />
       <div className="flex-1 flex flex-col min-w-0">
         <SuperAdminTopbar onMenuClick={() => setSidebarOpen(true)} />
         <main className="flex-1 p-4 lg:p-6 overflow-auto">
@@ -40,6 +42,11 @@ export default function SuperAdminLayout({ children }: { children: React.ReactNo
           </div>
         </main>
       </div>
+      <SupportBot
+        open={supportOpen}
+        onOpen={() => setSupportOpen(true)}
+        onClose={() => setSupportOpen(false)}
+      />
     </div>
   );
 }
